pyproject.toml
[project]
name = "umn-ta2-mineral-site-linkage"
version = "2.0.0"
description = "FuseMine: CriticalMAAS TA2 pipeline to link mineral site record"
authors = [
    {name = "Jiyoon Pyo",email = "jiyoonp0228@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars (>=1.19.0,<2.0.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "requests (>=2.32.3,<3.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "torchmetrics (>=1.6.1,<2.0.0)",
    "transformers (>=4.48.0,<5.0.0)",
    "datasets (>=3.2.0,<4.0.0)",
    "sentence-transformers (>=3.3.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
